allow HttpServer to run without Docile.
This is one major advantage of using bare strings for docstrings:
it doesn't break packages that don't yet support it.

I found that I needed to do this because I need fixes that have
happened in HttpServer since the Docile dependency was added but
the Docile framework is not stable enough to actually work every-
where. This is a hack to work around that catch 22.