Ah, sit down by the fire sonny and I'll tell you the story of a nasty bug.....
So HTTP.jl is a sneaky fella in that he tries a few tricks to detect if you're running a request asynchronously.
In the case the user IS, and the user has passed in a bonafide FIFOBuffer, he enters 'streaming-upload' mode.
This is great, because sometimes you want to do this.

In this case, he shot himself in the foot however, because although the user was running the request asynchronously,
there was no body to be found. HTTP.jl, however, was only detecting the asynchronous nature and so applied the 'chunked-transfer'
header, anticipating a streaming body upload. Later on, however, there was a check if there even is a body, and so any chunked
transfer was skipped, leaving us in a bad state, i.e. the remote host was still waiting for us to finish the request.
No wonder we kept timing out!!

That's all for today kids, come back next time to hear more tall tales of nasty http bugs.....

Fixes #18.