Make HandlerFunction type callable. e.g. (h::RequestHandlerFunction)(req) = handle(h,req)
This allows `listen` to treat the `h` handler parameter is a compeltely
generic way. i.e. functions, or Handlers or anything else that uses
the `object(args)` call pattern can be used as a request handler.

When `stream::Bool=true` the `::Stream` is passed to the handler `h` directly.
When `stream::Bool=false` the handler is wrapped a call to `handle(h, ::Stream)`.