Removed AWS4AuthRequest Layer
AWS4AuthRequest Layer lived in this package because previously you could
not insert your own custom layers into the HTTP stack. With the changes
in #443 we can now insert freely into the stack. We no longer need to
have AWS code living in this package and it can be removed. The AWS4Auth
layer now lives in the AWSCore.jl package.

- https://github.com/JuliaCloud/AWSCore.jl/commit/19e20b6t