Refactor connection pooling.
Instead of a single global connection pool, we now keep a connection
pool per "connection" (hash of distinguishing properties), also known as
a "pod". `getconnection` now checks for any valid, idle connections in
the pod channel, and will re-use if so. For any closed or invalid
connections (timed out, overused, etc.), they simply get dropped out of
the channel (and `numactive` of the pod is decremented).

If there are no valid connections to re-use from the pod channel, we
then check if the `numactive` connections in the pod is under the max;
if so, we create a new connection and increment `numactive`. If we've
hit our max # of allowed connections for a host:port, then we call
`take!(pod.channel)`, which will wait until a previous connection has a
transaction that calls `closewrite`, meaning it's available for re-use
or closed. Either way, we can check if the connection is valid and
re-use, or dump it and create a new one or wait again.

So it's important to note that _all_ connections are returned to the pod
channel when `closewrite` is called, and it can only be called once,
and _must_ be called after every `startwrite` regardless of any current
pipelining, request errors, the connection is closed unexpectedly, etc.

Another change was to hash on whether the connection was being used by a
client or server; unsurprisingly, when servers/clients using the same
connection pool try to pipeline requests/responses, things get messy. So
if someone is running a server/client both locally, we just make sure
their connection pool pods are separate.