2nd big set of changes. This time, we're removing the writedone/readdone Condition fields of the Connection object in favor of more threadsafe workflows. Specifically, we now represent the "readlock" and "writelock" of a Connection as a Channel{Bool} with size 1. Transactions will `take!` the single Bool element of the lock in order to "own" the lock while reading/writing. Once the transaction calls `closeread`/`closewrite`, they `put!` the "lock" back in the Connection object. This works well for pipelining, because a subsequent request can just wait by `take!`ing the lock it needs and it will become available as soon as the previous request finished and `put!`s the lock back.
For now, we leave the `sequence`, `writecount`, and `readcount`
variables as is in the Connection object. It might be better to have
these be `Threads.Atomic{Int}`, but in general, the Connection objects
are not super contentious, and these variables are only used for
incrementing Transactions, and checking for re-use/pipeline limiting.

Try to ensure that each pipelined request is executed in order