[BREAKING] Don't close response_stream after writing response body
Fixes #526. The original issue was when a response_stream was passed to
a request, and that request directed, the _redirect_ body was written
to the response_stream, then the response_stream was closed, then we
followed the redirect, and tried to write the subsequent _redirected_
body to the response_stream again, but it was already closed. Curiously,
`IOStream` doesn't throw a non-writeable error when trying to write to
it closed.

The change here involves two things:
* Not closing response_stream after we've written the response body to
it; this is the technically breaking piece, though I think it's probably
not super common that people were relying on this behavior. The HTTP.jl
tests had a few cases of using a `Base.BufferStream` that expected to be
closed, but even the HTTP.jl docs for response_stream include doing the
open-io-with-do-block example, which will close the IOStream for you.
* The 2nd change is that in the RedirectLayer, we now call `mark` and
`reset` on the response_stream, to ensure that the _redirected_ body
gets written where the original _redirect_ body was written to.

I'd appreciate any feedback/concerns on these changes; they sound like
good things in theory, but I'm just not sure if we'll be really breaking
anything for people.