Ensure https proxy connections don't clog up the ConnectionPool
Fixes #562. Despite lots of back and forth on this issue, it turns out
there's a simpler diagnosis & fix that worrying too much about when
sockets will throw when `close`ing (which we should certainly get all
cleaned up one day).

The _real_ issue here is in `ConnectionPool.sslupgrade` which was
effectively swapping our `Connection` that we formally "acquired" from
the connection pool with a new one that had a https-ssl-upgraded
`SSLContext` to do the request tunneling. The issue with just swapping
the `Connection` is that the original connection was never properly
"released" to the connection pool, so connections were acquired, but
never released, and eventually we hit the `connection_limit`. The
proposed solution here (which works locally to avoid getting hung up on
the connection pool clogging), is to introduce a new
`kill!(c::Connection)` method that "releases" a connection without
returning it to the connection pool. So we signal to the pool that a
connection has finished, but it's not put back in the queue to be
re-used, which is what we want. I should note that this does mean that
proxied https requests then don't really respect the `connection_limit`,
but also note that it never has due to this "swapping".