Fix readtimeout for requests
Should fix #571. There were a couple issues with the current
`readtimeout` and `TimeoutRequest` components. First, in the
`TimeoutRequest` layer, we were checking `inactiveseconds(io)` on the
`Transaction`, which returned 0 if the transaction was busy, which it
always would be in the middle of a request. Looking at uses of
`inactiveseconds`, it shouldn't be checking if the connection is busy or
not, so I removed that. The other places that were calling
`inactiveseconds` were already checking if the connection was busy or
not anyway.

The 2nd issue was that we weren't actually updating the connection
timestamp each time we read data for a response body. This would be
problematic for long-running downloads, because if if the request lasted
longer than the `readtimeout`, the connection would be closed, even
though we were still actively receiving data.

This should hopefully fix #571 since we'll now actually abort the
request if we reach the `readtimeout` instead of just sitting, waiting
indefinitely for lost packets of data to come. I will note that
currently `close(::SSLContext)` doesn't "interupt" the connection like
it does for `TcpSocket`, so the behavior is slightly different. But it's
a much larger (and scarier) project to overhaul the `SSLContext` `close`
behavior/semantics (but still needs to be done at some point). Hence
only the reliable test for `http` for now.