Catch getsslcontext errors in server listenloop
Fixes #318. This is a bad bug. Basically, if you're running an ssl
server, any client who sends a non-perfect https request (like simply a
plain *http* request) would result in the server task throwing an error
and closing. This is partly because the thrown error isn't super
obvious, nested down a few calls in our overloaded
`Sockets.accept(::Server{SSLConfig})` method. The fix proposed here is
that we'll put a try-catch in the ssl handshaking code and if an error
is thrown, we return `nothing`; in the server `listenloop`, if the
accepted `io` is `nothing`, then we'll skip and try to accept the next
connection.