Flush TCP connections during `closewrite()`
TCP connections are generally (depending on kernel whimsy) governed by
the Nagle algorithm, which may not immediately transmit data but instead
wait for the TCP buffer to fill before spitting a packet out onto the
network.  This is due to the fact that the kernel does not necessarily
know when a "message" is complete and simply does its best to maximize
throughput (by decrease packet counts) while minimizing latency (by
transmitting quickly).  We can help the OS out a bit here by flushing
the TCP buffer when we're done with a message (e.g. in `closewrite()`)
which (on some kernels) reduces the minimum latency of a message from
~40ms to < 1ms.  We flush by toggling the Nagle algorithm off and then
on again on that particular socket, forcing the OS to transmit any
pieces of data that were queued to be sent out of that socket
immediately.