Handle poorly-behaved callbacks more reliably
If `startwrite()` is never called during a server callback, the client
may be left waiting for a response and never get one.  This can cause
lockups and general misery, especially when connections are proxied
through an intermediary (such as `nginx`) with a finite connection
limit.

This PR detects a server callback that never calls `startwrite(http)`
and throws an error, triggering the HTTP 500 mechanism.  It also adds a
`closewrite(http)` onto that branch as without it, it appears that the
final chunk in the `Transfer-Encoding: Chunked` response may not appear.