Change `connect_timeout` to `connecttimeout` and document
This is to make it the same format as `readtimeout`.

Also changes `connecttimeout` to be the nonzero minimum of
`connecttimeout` and `readtimeout` if either is set.

This is to get around connections hanging even if
`readtimeout` is set. For example, this occurs if you try
to connect to the EC2 metadata IP from outside AWS.