Store remote IP adress in connection struct
The remote IP is used in access logging but if the connection is closed
at the time of creating the log message it is not possible to fetch it.
(In fact, it currently segfaults, but will error in future Julia versions,
see JuliaLang/julia#40993). With this patch the remote IP is stored in
the Connection struct. This should not cost anything extras since the call
to Sockets.getpeername(...) is already there for storing the remote port.