Do not call `close()` on the `response_stream`
This allows a plain `IOBuffer` to be used with `response_stream`.

It seems unnecessary to be calling `close(response_stream)` inside
HTTP.request() - given that HTTP.request is a blocking operation, the
caller can easily close the stream themselves after `HTTP.request()`
returns.

This could be considered a breaking change for users who use the
internal IO type Base.BufferStream with the response_stream keyword.

Fixes #543