Two changes to fix windows tests (#802)
* Two changes to fix windows tests

The first change is in our `Connection` `unsafe_read` method where we
wrap our `unsafe_read` call to the underlying socket in a try-catch and
if an `IOError` is thrown, just rethrow as `EOFError`. The reasoning
here is left as a code comment, but in short, at the `Connection` level,
we don't really care whether the underlying socket throws an `EOFError`
or the socket received an RST or whatever, we really just need to know
if *something* went wrong and bubble that up to the header/body parsing
code in a consistent way.

The 2nd change is around some failing server tests; the problem with
sending a request with larger-than-allowed headers is the server
immediately bails on parsing the headers (to not DOS itself) and
immediately closes the underlying socket. The closing of the socket in
this abrupt manner means the client may or may not actually recieve the
413 response, which as the server we don't really care since they were
probably sending the request maliciously anyway. So I'm not exactly sure
the best way to encode that in the testing framework, but for now, I'm
kind of marking them as "allowed to fail".

* additional fixes

* Remove deprecated tests that failed on windows