Attempt to reencode malformed headers from Latin-1 to UTF8 (#830)
* Attempt to reencode malformed headers from Latin-1 to UTF8

As brought up in #796, there may be scenarios where headers may contain
non-UTF8 characters (even though they're supposed to be ASCII). Appreciation
to @StefanKarpinski for the Latin-1 -> UTF-8 conversion code and the suggestion
to try reencoding before throwing an error. As proposed in this PR, the normal
header parsing path should be unaffected and only when we're unable to parse
a normal header will we attempt this reencoding.

Note that curl warns on the malformed header and filters it out.

* cleanup implementation and ensure tests pass

* fix

* one more cleanup