Auto decompress gzip-encoding response bodies (#838)
* Auto decompress gzip-encoding response bodies

Implements #256. If the content-encoding of a response is "gzip"
and the keyword argument `decompress === true`, then we'll
use CodecZlib.jl to decompress the response and set as the response
body. Passing `decompress=false` will leave the resposne body as-is.
We also support `HTTP.decode(::Request, "gzip")` which will do
the decompression.

* fixes

* fix

* fix

* fix'

* fix