Alternative approach for response streams since mark/reset have issues, namely we may write a 500 response body to the response_stream, reset it to original starting position, retry the request, and then not get as many bytes in the successful response body as we got in the error case, which leads to the response_stream having the successful body followed by the rest of the error body.
Instead, this commit proposes we take a similar approach to how we handle
redirects: namely we avoid writing to the response_stream unless we know
we won't be redirecting, or in this new case, retrying anymore. I cleaned
up the redirect approach by utilizing the new request.context store,
which makes it easier to flag when we've reached the limit for redirecting
or retrying without having to pass state around as keyword args.