Ensure underlying Connection gets closed in websocket close sequence if timer runs out
Noticed this while doing a local back and forth w/ a manual server socket.
The problem is when we try to initiate the websocket close sequence, but we
have a remote who just won't respond (like in my manual socket case). I
*thought* the Timer would kick in and close things up, but it turns out
the `receive(ws)` call will just block indefinitely, so we never manage
to break out of our while loop. The fix proposed here adds the logic to
the Timer, so if we wait 5 seconds and nothing has been received, we
actually go ahead and close the underlying Connection which in turn
will break us out of the `receive` while loop and not cause `close(ws)`
to hang.