Store the original registered route path when Router matches (#866)
* Store the original registered route path when Router matches

This is a common feature of other router implementations, sometimes referred
to as "reverse routing". We don't go the full way in allowing you to do
`getregisteredroute(r::Router, url)`, but we could do that eventually. For now,
it's at least very useful to store the registered route that was matched
during request handling.

* don't overload base methods

* fix