Cleanup acquire/release usage from underlying connection pool
From issues reported on slack, I was able to reproduce a hang
when trying to concurrently download thousands of files. The issue
is that there's a small window from when a Connection was acquired
from the connection pool to when we were finished reading from the connection
that if an error is thrown, it wasn't getting properly released back
to the connection pool. So if you're doing thousands of requests,
that's bound to happen and as the connection pool fills up with these
unreleased connections, that leaves less and less room for real active
connections, eventually even getting completely full and "hanging".

This PR proposes a fix where we explicitly release the connection
via try-finally block right after we've acquired it, to ensure
each connection gets released, regardless of any other errors
that may happen.