Add ability to "hook" a middleware into Router post-matching (#886)
This is a common feature of other router/handler frameworks. I initially
didn't know if we'd need it, but I realized the convenient use-case is
when you want a "hook" _after_ you've matched the route, but before you
dispatch to the matched handler. This can be nice for argument processing
middleware or metric logging on matched routes.