Cleanup all timeout code and places where we might timeout
Fixes #909. Great issue that stretches our current system and exposes
a bunch of places where we weren't accounting for timeouts. The core
issues here are when we have a proxy involved, we had a couple more
places where we were connecting/writing/reading, but weren't applying
a user-provided `connect_timeout` or `readtimeout`. Also for our
ssl upgrade path. In the OP's example, we're hitting thousands of mostly
non-functional proxies, so it's basically a huge stress test on all the
error paths of our request code. This PR proposes:
  * a new `try_with_timeout` function to provide a consistent way to run
    code while waiting for a timeout
  * use the new function in a number of places where we might be stuck waiting

I'm able to get the original script OP provided to finish in ~150 seconds, and
I tried out a multithreaded version that finished in about ~17 seconds w/ 8 threads.