Allow retrying request regardless of idempotent if we failed to write… (#964)
* Allow retrying request regardless of idempotent if we failed to write anything

There isn't one specific issue this addresses, but I've received reports
from a few private sources around production issues people have run into
with a `IOError: NET - Connection was reset by peer` error with a stacktrace
coming from https://github.com/JuliaWeb/HTTP.jl/blob/37fec3578428f4033c5a031d9d418fb96edbffb9/src/Streams.jl#L81,
which is the 1st place we try to write the request start line + headers on the wire.

After investigating the MbedTLS.jl side of things (this happens with https connections),
from what I can tell, the core issue is a race condition between the remote server closing
a connection we just used in a previous request and when we try to reuse that request
in a subsequent, _non-idempotent_ request (i.e. POST). So the problem manifests with
the following sequence of events:
  * We make a request, it completes successfully
  * We try to make a subsequent, non-idempotent request and reuse the previous request connection
  * When we initially retrieve the connection from the connection pool, it looks open/valid
  * In between the time we retrieve the conn from the pool and the time we _actually_ start
    writing our non-idempotent request, the remote closes the connection
  * Subsequently, our first write fails because the connection is closed
  * Our retry logic _doesn't_ attempt to retry because the request is idempotent

This theory is partly confirmed by reports that this issue can be "solved" by calling
`HTTP.ConnectionPool.closeall()` before the non-idempotent request, which forces a
brand-new connection to be made for the request.

This PR proposes the following solution:
  * When we create a `Stream`, we set the `nwritten` field to -1 to signal
    that we haven't written the start line or headers
  * `nwritten` is set to 0 once we've finished writing request headers
  * In `ConnectionRequest`, when we catch a downstream error, we set a
    `nothingwritten` values in the `Request` `context` dict to signal
    that nothing was written for this request yet
  * In the `retryable(::Request)` check, we now check if the request is
    idempotent, the user passed `retry_non_idempotent` _OR_ (new) if we failed
    to write anything in the first place

* overhaul tests to httpbingo.julialang.org for better reliability