Fix access_threaded in the forthcoming interactive threadpool world
Fixes #970.

The issue here is that our `access_threaded` function, and thread-specific
arrays, all relied on `Threads.nthreads()`, which by default only returns
the number of threads in the `:default` threadpool. In the new world where
the `:interactive` threadpool exists, there can also be threads there.

There is a new function `Threads.maxthreadid()` which helpfully tells us
the max id of any thread, regardless of threadpool. So the proposed fix
here is that if that function is defined, use that, otherwise fallback
to `Threads.nthreads()`.