Make timeouts default to 60 seconds
The client-side timeout implementation has gotten much more reliable
and robust over the last little while. The reason we didn't include
a timeout by default is that it wasn't in the past, but now I
believe things are robust enough to set it and catch unresponsive
servers for clients without waiting on the response indefinitely.