Fixes EACCESS while running package tests
Some of the tests make use of docker containers with mounted volumes fromhost filesystem. That causes files to be created as root and attempt to delete them results in EACCESS.

This changes the test code in a couple of places to use the same docker containers to delete the files that they create.

fixes #1086