Make CI green again (#1128) (#1152)
* Fix tests for new version of `ConcurrentUtilities`

`ConcurrentUtilities` changed an (internal) fieldname of the `Pool` type
from `max` to `limit`. This field is used in HTTP.jl tests so this patch
works around the issue.

* Remove invalidation CI check

This has been failing for quite some time with some internal error. Also
in the working case it is a bit unclear what the actionable thing is
when it errors (xref
https://github.com/JuliaWeb/HTTP.jl/pull/920#issuecomment-1254284159).

Co-authored-by: Fredrik Ekre <ekrefredrik@gmail.com>